chore(e2e): move navigation utils to their own file
