+++ e2e/src/usecases/NewAccountOnboarding.js
@@ -1,12 +1,7 @@
 import { sleep } from '../../../src/utils/sleep'
 import { launchApp } from '../utils/retries'
-import {
-  completeProtectWalletScreen,
-  enterPinUi,
-  navigateToSecurity,
-  quickOnboarding,
-  waitForElementById,
-} from '../utils/utils'
+import { navigateToSecurity } from '../utils/navigation'
+import { enterPinUi, quickOnboarding, waitForElementById } from '../utils/utils'
 
 import jestExpect from 'expect'
 
@@ -29,6 +24,14 @@ const startBackupFromNotifications = async () => {
 const arriveAtHomeScreen = async () => {
   // Arrived to Home screen
   await waitForElementById('HomeAction-Send')
+}
+
+const completeProtectWalletScreen = async () => {
+  await expect(element(by.id('recoveryPhraseCard'))).toBeVisible()
+  await element(by.id('recoveryPhraseCard')).tap()
+  await enterPinUi()
+  await expect(element(by.id('protectWalletBottomSheetContinue'))).toBeVisible()
+  await element(by.id('protectWalletBottomSheetContinue')).tap()
 }
 
 export default NewAccountOnboarding = () => {
