+++ e2e/src/usecases/PINChange.js
@@ -1,7 +1,8 @@
 import { sleep } from '../../../src/utils/sleep'
 import { ALTERNATIVE_PIN, DEFAULT_PIN } from '../utils/consts'
 import { reloadReactNative } from '../utils/retries'
-import { enterPinUi, navigateToSecurity, waitForElementById } from '../utils/utils'
+import { enterPinUi, waitForElementById } from '../utils/utils'
+import { navigateToSecurity } from '../utils/navigation'
 
 export default ChangePIN = () => {
   it('Then should be retain changed PIN', async () => {
